Seats and moviepoints methods, also ticket by user id.

+++ backend/Ticketcine/src/main/java/com/s1511/ticketcine/application/dto/user/ReadDtoUser.java
@@ -5,6 +5,6 @@ public record ReadDtoUser(
         String email,
         String firstName,
         String lastName,
-        Long moviePoints
+        Double moviePoints
 ) {
 }
